+++ src/main/java/com/sudurukbackback/modulecture/view/CommunityController.java
@@ -15,6 +15,11 @@ public class CommunityController {
         return "domain/community/community";
     }
 
+    @GetMapping("/post-detail")
+    public String postDetail(Model model) {
+        return "domain/community/post-detail";
+    }
+
     @GetMapping("/post-write")
     public String postWrite(Model model) {
         return "domain/community/post-write";
