+++ src/main/java/com/sudurukbackback/modulecture/view/MainController.java
@@ -12,35 +12,4 @@ public class MainController {
         return "domain/main";
     }
 
-    @GetMapping("/mypage")
-    public String mypage(Model model) {
-        return "domain/mypage/mypage";
-    }
-
-    @GetMapping("/mypage-lecture")
-    public String mypageLecture(Model model) {
-        return "domain/mypage/mypage-lecture";
-    }
-
-    @GetMapping("/mypage-tutor")
-    public String mypageTutor(Model model) {
-        return "domain/mypage/mypage-tutor";
-    }
-
-    @GetMapping("/mypage-tutor-upload")
-    public String mypageTutorUpload(Model model) {
-        return "domain/mypage/mypage-tutor-upload";
-    }
-
-    @GetMapping("/mypage-tutor-update")
-    public String mypageTutorUpdate(Model model) {
-        return "domain/mypage/mypage-tutor-update";
-    }
-
-    @GetMapping("/mypage-account")
-    public String mypageAccount(Model model) {
-        return "domain/mypage/mypage-account";
-    }
-
-
 }
