+++ src/main/resources/templates/domain/community/community.html
@@ -38,7 +38,7 @@
                             <tbody>
                             <!-- 공지사항 게시글 먼저 표시 -->
                             <!-- 테스트 데이터 행 1 (공지사항) -->
-                            <tr onclick="window.location='/post/1';" style="cursor: pointer;">
+                            <tr onclick="window.location='/community/post-detail';" style="cursor: pointer;">
                                 <td>
                                     <span class="badge bg-primary bg-opacity-75 text-align-top ms-3">공지사항</span>
                                 </td>
