+++ src/main/resources/templates/domain/community/post-detail.html
@@ -1,0 +1,137 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org" lang="ko">
+
+<!-- head 부분 포함 -->
+<head th:fragment="head" xmlns:th="http://www.thymeleaf.org">
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
+    <meta name="description" content="" />
+    <meta name="author" content="" />
+    <title>모두의강의 - 커뮤니티 </title>
+
+    <!-- CSS-CDN -->
+    <div th:replace="layout/css-cdn :: css-cdn"></div>
+</head>
+
+<body class="d-flex flex-column min-vh-100">
+
+<!-- navbar 부분 포함 -->
+<div th:insert="layout/navbar :: navbar"></div>
+
+<!-- Section -->
+<main class="flex-fill">
+    <section class="py-5">
+        <div class="container px-4 px-lg-5">
+            <div class="row">
+
+                <!-- mypage vertical navigation bar -->
+                <div th:replace="layout/community-navbar :: community-navbar"></div>
+
+                <!-- 오른쪽 내용 -->
+                <div class="col-md-9">
+<!--                    <h2 class="mb-4 mt-4 me-4">글 상세 보기</h2>-->
+<!--                    <hr>-->
+
+                    <!-- 게시글 내용 박스 -->
+                    <div class="border rounded p-4">
+                        <!-- 게시글 제목 -->
+                        <div class="mb-3 d-flex align-items-center">
+                            <h3>첫 번째 게시글</h3>
+                            <span class="badge bg-primary d-block mb-2 ms-3">공지사항</span>
+                        </div>
+                        <hr>
+
+                        <!-- 카테고리 및 작성자 정보 -->
+                        <div class="mb-3">
+                            <span class="text-muted">작성자: <span>관리자</span></span>
+                        </div>
+
+                        <!-- 게시글 내용 -->
+                        <div class="mb-5">
+                            <p>이곳에 게시글의 내용이 표시됩니다. 예시로 간단한 내용을 작성하였습니다.</p>
+                        </div>
+                    </div>
+
+                    <!-- 댓글 섹션 시작 -->
+                    <div class="mt-4">
+                        <!-- 댓글 개수 및 댓글 버튼 -->
+                        <div class="d-flex justify-content-between align-items-center mb-3">
+                            <h5>
+                                <i class="bi bi-chat"></i> 2개
+                            </h5>
+                            <button id="commentButton" class="btn btn-primary">댓글 작성</button>
+                        </div>
+
+                        <hr>
+
+                        <!-- 기존 댓글 목록 -->
+                        <div id="commentList" class="mb-4">
+                            <!-- 댓글 1 -->
+                            <div class="mb-3">
+                                <div class="d-flex justify-content-between">
+                                    <strong>사용자1</strong>
+                                    <span class="text-muted">2024-03-01</span>
+                                </div>
+                                <p>첫 번째 댓글 내용입니다.</p>
+                                <hr>
+                            </div>
+                            <!-- 댓글 2 -->
+                            <div class="mb-3">
+                                <div class="d-flex justify-content-between">
+                                    <strong>사용자2</strong>
+                                    <span class="text-muted">2024-03-01</span>
+                                </div>
+                                <p>두 번째 댓글 내용입니다.</p>
+                                <hr>
+                            </div>
+                        </div>
+
+                        <!-- 댓글 작성 폼 -->
+                        <div id="commentForm" style="display: none;">
+                            <div class="mb-3">
+                                <textarea class="form-control" id="commentContent" name="content" rows="3"></textarea>
+                            </div>
+                            <!-- 댓글 등록 버튼을 오른쪽으로 정렬 -->
+                            <div class="text-end">
+                                <button type="button" id="submitComment" class="btn btn-primary">댓글 등록</button>
+                            </div>
+                            <hr>
+                        </div>
+                    </div>
+                    <!-- 댓글 섹션 끝 -->
+                </div>
+
+            </div>
+        </div>
+    </section>
+</main>
+
+<!-- Footer -->
+<div th:insert="layout/footer :: footer"></div>
+
+<!-- JS CDN -->
+<div th:replace="layout/js-cdn :: js-cdn"></div>
+
+<!-- 댓글 작성 버튼 클릭 시 폼 표시하는 스크립트 -->
+<script>
+    document.getElementById('commentButton').addEventListener('click', function() {
+        // 댓글 작성 폼 표시
+        document.getElementById('commentForm').style.display = 'block';
+        // 댓글 작성 버튼 숨기기
+        document.getElementById('commentButton').style.display = 'none';
+    });
+
+    // 댓글 등록 버튼 클릭 시 처리 (실제 구현에서는 Ajax 등의 방법으로 서버에 전송해야 합니다)
+    document.getElementById('submitComment').addEventListener('click', function() {
+        // 예시로 alert 표시
+        alert('댓글이 등록되었습니다.');
+        // 댓글 등록 후 폼 숨기기 및 내용 초기화
+        document.getElementById('commentForm').style.display = 'none';
+        document.getElementById('commentContent').value = '';
+        // 댓글 작성 버튼 다시 표시
+        document.getElementById('commentButton').style.display = 'inline-block';
+    });
+</script>
+
+</body>
+</html>
