+++ src/main/resources/templates/domain/mypage/mypage-tutor.html
@@ -30,7 +30,7 @@
                 <div class="d-flex justify-content-between align-items-center mt-3 mb-4">
                     <h2>나의 강의 관리</h2>
                     <!-- 강의 등록 버튼 -->
-                    <a href="/mypage-tutor-upload" class="btn btn-primary">강의 등록</a>
+                    <a href="/mypage/tutor-upload" class="btn btn-primary">강의 등록</a>
                 </div>
 
                 <hr>
@@ -53,7 +53,7 @@
                                 <h6 class="text-dark text-opacity-75">김모두 강사</h6>
                                 <hr>
                                 <p class="card-text">강의 소개 내용 1입니다.</p>
-                                <a href="/mypage-tutor-update" class="btn btn-primary">수정하기</a>
+                                <a href="/mypage/tutor-update" class="btn btn-primary">수정하기</a>
                             </div>
                         </div>
                     </div>
