+++ src/main/resources/templates/layout/community-navbar.html
@@ -5,13 +5,10 @@
         <h4 class="mb-4 text-start">커뮤니티</h4>
         <ul class="nav flex-column text-start">
             <li class="nav-item">
-                <a class="nav-link ps-3 menu-link border-start text-dark hover-bg-blue" href="/mypage">전체 글</a>
+                <a class="nav-link ps-3 menu-link border-start text-dark hover-bg-blue" href="/community">전체 글 보기</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link ps-3 menu-link border-start text-dark" href="/mypage-lecture">나의 글</a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link ps-3 menu-link border-start text-dark" href="/mypage-tutor">나의 댓글</a>
+                <a class="nav-link ps-3 menu-link border-start text-dark" href="/community/post-mine">나의 글 보기</a>
             </li>
             <br>
             <!-- 글쓰기 버튼 -->
