+++ src/main/resources/templates/layout/mypage-navbar.html
@@ -8,13 +8,13 @@
                 <a class="nav-link ps-3 menu-link border-start text-dark hover-bg-blue" href="/mypage">나의 정보</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link ps-3 menu-link border-start text-dark" href="/mypage-lecture">수강 목록</a>
+                <a class="nav-link ps-3 menu-link border-start text-dark" href="/mypage/lecture">수강 목록</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link ps-3 menu-link border-start text-dark" href="/mypage-tutor">나의 강의 관리</a>
+                <a class="nav-link ps-3 menu-link border-start text-dark" href="/mypage/tutor">나의 강의 관리</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link ps-3 menu-link border-start text-dark" href="/mypage-account">계정</a>
+                <a class="nav-link ps-3 menu-link border-start text-dark" href="/mypage/account">계정</a>
             </li>
         </ul>
     </div>
