+++ src/main/resources/templates/layout/navbar.html
@@ -2,7 +2,7 @@
 <nav th:fragment="navbar" class="navbar navbar-expand-lg navbar-light bg-light shadow">
     <div class="container px-4 px-lg-5">
         <a class="navbar-brand" href="/main">
-            <img src="images/logo.svg" alt="Logo">
+            <img src="/images/logo.svg" alt="Logo">
         </a>
         <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
                 data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
@@ -18,8 +18,8 @@
                        data-bs-toggle="dropdown" aria-expanded="false">마이페이지</a>
                     <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
                         <li><a class="dropdown-item" href="/mypage">나의 정보</a></li>
-                        <li><a class="dropdown-item" href="/mypage-lecture">수강 목록</a></li>
-                        <li><a class="dropdown-item" href="/mypage-tutor">나의 강의 관리</a></li>
+                        <li><a class="dropdown-item" href="/mypage/lecture">수강 목록</a></li>
+                        <li><a class="dropdown-item" href="/mypage/tutor">나의 강의 관리</a></li>
                     </ul>
                 </li>
             </ul>
